FOGL-9486 Fix issue on armv7l platforms

+++ C/plugins/storage/postgres/connection.cpp
@@ -1130,7 +1130,7 @@ SQLBuffer	sql;
 					else if (itr->value.IsDouble())
 						sql.append(itr->value.GetDouble());
 					else if (itr->value.IsUint64())
-						sql.append(itr->value.GetUint64());
+						sql.append((unsigned long)itr->value.GetUint64());
 					else if (itr->value.IsInt64())
 						sql.append(itr->value.GetInt64());
 					else if (itr->value.IsObject())
